Acl finished

+++ app/controllers/ControllerBase.php
@@ -2,6 +2,9 @@
 
 use Phalcon\Mvc\Controller;
 use Phalcon\Acl\Adapter\Memory as AclList;
+use Phalcon\Acl\Role;
+use Phalcon\Acl\Resource;
+use Phalcon\Mvc\Url;
 
 class ControllerBase extends Controller{
 
@@ -17,9 +20,13 @@ class ControllerBase extends Controller{
     }
 
     public function indexAction(){
-        $objects = call_user_func($this->model."::find");
-        $this->view->setVar("objects",$objects);
-        $this->view->pick("main/index");
+    	if ($this->verifyAccessAction($this->controller, "index")) {
+	        $objects = call_user_func($this->model."::find");
+	        $this->view->setVar("objects",$objects);
+	        $this->view->pick("main/index");
+    	} else {
+    		echo $this->jquery->bootstrap()->htmlAlert("alert1","Vous n'avez pas les droits pour accéder à cette page !");
+    	}
     }
 
     public function frmAction($id =  NULL){
@@ -35,26 +42,48 @@ class ControllerBase extends Controller{
     }
 
     public function updateAction(){
-        echo "Pas encore implémenté...";
+    	$id = $this->request->getPost('id','int');
+        if ($this->request->isPost()) {
+            $object = $this->getInstance(@$_POST["id"]);
+            $this->setValuesToObject($object);
+            if ($id) {
+                try {
+                    $object->save();
+                    $msg = new DisplayedMessage("Instance de " . $this->model . " modifiée");
+                } catch (\Exception $e) {
+                    $msg = new DisplayedMessage("Impossible d'ajouter l'instance de " . $this->model, "danger : $e");
+                }
+            } else {
+                try {
+                    $object->save();
+                    $msg = new DisplayedMessage("Instance de " . $this->model . " ajoutée");
+                } catch (\Exception $e) {
+                    $msg = new DisplayedMessage("Impossible d'ajouter l'instance de " . $this->model, "danger : $e");
+                }
+            }
+            $this->dispatcher->forward(array("controller" => $this->dispatcher->getControllerName(), "action" => "index", "params" => array($msg)));
+        }
     }
 
     public function deleteAction($id = null){
-        $object = call_user_func($this->model.'::findFirst', "id = $id");
-        
-        $object->delete();
-        $this->response->redirect("Index/index");
+    	if ($this->verifyAccessAction($this->controller, "write")) {
+	        $object = call_user_func($this->model.'::findFirst', "id = $id");
+	        
+	        $object->delete();
+	        $this->response->redirect("Index/index");
+    	} else {
+    		echo "Access Denied !";
+    	}
     }
 
     public function asAdminAction(){
-        $user = new User();
-        $user->findFirst("idTypeUser = '0' ");
+        $user = User::findFirst("id = 3");
         $this->session->set("user", $user);
         $this->response->redirect("Index/index");
     }
 
     public function asUserAction(){
-        $user = new User();
-        $user->findFirst("idTypeUser = '2' ");
+        $user = User::findFirst("id = 5");
         $this->session->set("user", $user);
         $this->response->redirect("Index/index");
     }
@@ -63,4 +92,47 @@ class ControllerBase extends Controller{
         $this->session->destroy();
         $this->response->redirect("Index/index");
     }    
+    
+    public function loadAclAction($typeUser) {
+    	$acl = new AclList();
+    	$acl->setDefaultAction(Phalcon\Acl::DENY);
+    	
+    	$roles = TypeUser::find();
+    	foreach ($roles as $role) {
+    		$acl->addRole($role->getLibelle());
+    	}
+
+    	$operationsBdd = Operation::find();
+    	$operations = array();
+    	foreach ($operationsBdd as $operation) {
+    		$operations[] = $operation->getOperation();
+    	}
+    	
+    	$ressources = Ressource::find();
+    	foreach ($ressources as $ressource) {
+    		$acl->addResource($ressource->getLibelle(), $operations);
+    	}
+ 
+    	$aclsBdd = Acl::find();
+    	foreach ($aclsBdd as $aclBdd) {
+    		$typeUserBdd = TypeUser::findFirst("id = ".$aclBdd->getIdTypeUser());
+    		$ressourceBdd = Ressource::findFirst("id = ".$aclBdd->getIdRessource());
+    		$operationBdd = Operation::findFirst("id = ".$aclBdd->getIdOperation());
+    		$acl->allow($typeUserBdd->getLibelle(), $ressourceBdd->getLibelle(), $operationBdd->getOperation());
+    	}
+    	
+    	return $acl;
+    }
+    
+    public function verifyAccessAction($activeResource, $activeOperation) {
+    	$user = $this->session->get("user");
+    	$typeUser = TypeUser::findFirst("id = ".$user->getIdTypeUser());
+    	$typeUserSession = $user->getIdTypeUser();
+    	$acl = $this->loadAclAction($typeUserSession);
+    	if ($acl->isAllowed($typeUser->getLibelle(), $activeResource, $activeOperation)) {
+    		return 1;
+    	} else {
+    		return 0;
+    	}
+    }
 }
